+++ apps/files_external/lib/config/configadapter.php
@@ -34,7 +34,7 @@ use OCP\IUser;
 use OCA\Files_external\Service\UserStoragesService;
 use OCA\Files_External\Service\UserGlobalStoragesService;
 use OCA\Files_External\Lib\StorageConfig;
-use OCA\Files_External\Lib\FailedStorage;
+use OC\Files\Storage\FailedStorage;
 use OCP\Files\StorageNotAvailableException;
 
 /**
