+++ lib/private/files/cache/failedcache.php
@@ -19,9 +19,8 @@
  *
  */
 
-namespace OCA\Files_External\Lib;
+namespace OC\Files\Cache;
 
-use OC\Files\Cache\CacheEntry;
 use OCP\Constants;
 use OCP\Files\Cache\ICache;
 
@@ -29,6 +28,7 @@ use OCP\Files\Cache\ICache;
  * Storage placeholder to represent a missing precondition, storage unavailable
  */
 class FailedCache implements ICache {
+	/** @var bool whether to show the failed storage in the ui */
 	private $visible;
 
 	/**
